refactor subtitle download

+++ src/main/helpers.ts
@@ -1,5 +1,6 @@
 import fs from "fs";
 import https from "https";
+import path from "path";
 
 export const generatePortNumber = () => {
   const min = 1024;
@@ -12,16 +13,41 @@ export const generatePortNumber = () => {
 export const toUnixPath = (path: string) =>
   path.replace(/[\\/]+/g, "/").replace(/^([a-zA-Z]+:|\.\/)/, "");
 
-export const download = (url: string, dest: string) => {
+export const download = (
+  url: string,
+  dir: string,
+  filename?: string
+): Promise<string> => {
   return new Promise((resolve, reject) => {
-    const file = fs.createWriteStream(dest, { flags: "wx" });
+    const regexp = /filename=\"(.*)\"/gi;
 
     const request = https.get(url, (response) => {
       if (response.statusCode === 200) {
+        filename =
+          filename ||
+          (regexp.exec(response.headers["content-disposition"]) || [])[1];
+        filename = filename || new URL(url).pathname.split("/").pop();
+        const dest = path.join(dir, filename);
+        const file = fs.createWriteStream(dest, {
+          flags: "wx",
+        });
         response.pipe(file);
+
+        file.on("finish", () => {
+          resolve(dest);
+        });
+
+        file.on("error", (err: any) => {
+          file.close();
+
+          if (err.code === "EEXIST") {
+            resolve(dest);
+          } else {
+            fs.unlink(dest, () => {}); // Delete temp file
+            reject(err.message);
+          }
+        });
       } else {
-        file.close();
-        fs.unlink(dest, () => {}); // Delete temp file
         reject(
           `Server responded with ${response.statusCode}: ${response.statusMessage}`
         );
@@ -29,24 +55,7 @@ export const download = (url: string, dest: string) => {
     });
 
     request.on("error", (err) => {
-      file.close();
-      fs.unlink(dest, () => {}); // Delete temp file
       reject(err.message);
-    });
-
-    file.on("finish", () => {
-      resolve(dest);
-    });
-
-    file.on("error", (err: any) => {
-      file.close();
-
-      if (err.code === "EEXIST") {
-        resolve(dest);
-      } else {
-        fs.unlink(dest, () => {}); // Delete temp file
-        reject(err.message);
-      }
     });
   });
 };
